requirements.txt
google-adk>=1.2.0
mcp[cli]>=1.0.0
httpx>=0.27.0
